Initial commit.
The source code is kind of a mess. I've been converting it from an
actual AR interface, into the program that is going to run for our EOG
tests, while trying to minimise the changes. This is not working; hence
I am creating this repo, so I can develop the testing program on a
separate branch. I should have done this ages ago.